Properly sync configuration to the branch projects.
Due to limitations with copying configuration directly from the parent
to the sub-projects, a template sub-project (separate from the branches)
is required to store the configuration in an XML format.  The XML is
then unmarshaled into each of the sub-projects.

Additional changes to the configuration jelly now point the fields to
the template project.

The FreeStyleBranchProject must also now include a true/false flag to
mark it as the template so that special logic can be used to store the
template in a location separate from the branches.  This is not possible
without using the flag, as the template sub-project will get loaded into
the branches directory when Jenkins is restarted.

Add additional logging for the sync branches log.

Complete saving of remaining fields to template & branches.