Refactor AbstractMultiBranchProject to inherit from AbstractItem.
Because the multi-branch projects themselves are just parents for
the branch projects and are not runnable themselves, it makes
more sense for them to be items.

This should remove some of the hacking around the AbstractProject
and Job APIs at the cost of adding some code to fill the gaps.

Simplified use of the SyncBranchesTrigger.  Added an additional
Cron extension for the trigger since standard Cron only checks
triggers on AbstractProjects.

Added APIs for last build, last successful build, etc. for view
columns.