Fix incompatiblity with Python program
If branch name contains "/", it would be encoded to "%2F" in current
implementation. But as for Python "%2F" is a presetation type when presented
in a string: it stands a floating point number to be formated.

Many Python based program cannot work well with this plugin if branch name
contains "/"

So this fix tries to replace "/" with "-" to avoid many potential failure.